re-export more
